requirement.txt
absl-py==2.1.0
aioredis==1.3.1
asgiref==3.4.1
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.2.0
autobahn==23.1.2
Automat==22.10.0
beepy==1.0.7
cachetools==5.3.2
certifi @ file:///C:/b/abs_85o_6fm0se/croot/certifi_1671487778835/work/certifi
cffi==1.15.1
channels==3.0.3
channels-redis==3.3.0
charset-normalizer==3.3.2
colorama==0.4.6
colorlog==6.7.0
constantly==15.1.0
cryptography==42.0.7
cvlib==0.2.7
cycler==0.11.0
Cython==3.0.6
daphne==3.0.2
Django==3.2.25
flatbuffers==23.5.26
fonttools==4.38.0
gast==0.4.0
google-auth==2.23.4
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.59.3
h5py==3.8.0
hiredis==2.0.0
hyperlink==21.0.0
idna==3.6
imageio==2.31.2
importlib-metadata==6.7.0
imutils==0.5.4
incremental==22.10.0
keras==2.11.0
kiwisolver==1.4.5
libclang==16.0.6
Markdown==3.4.4
MarkupSafe==2.1.3
matplotlib==3.5.3
mediapipe==0.8.9.1
MouseInfo==0.1.3
msgpack==1.0.5
networkx==2.6.3
numpy==1.21.6
oauthlib==3.2.2
opencv-python==3.4.2.16
opt-einsum==3.3.0
packaging==23.2
pandas==1.3.5
Pillow==9.5.0
progressbar==2.5
protobuf==3.19.6
pyasn1==0.5.1
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pynput==1.7.6
pyOpenSSL==24.1.0
pyparsing==3.1.1
pyperclip==1.8.2
PyRect==0.2.0
PyScreeze==0.1.30
python-dateutil==2.9.0.post0
pytweening==1.0.7
pytz==2023.3.post1
PyWavelets==1.3.0
redis==5.0.4
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.19.3
scipy==1.7.3
service-identity==21.1.0
simpleaudio==1.0.4
six==1.16.0
sqlparse==0.4.4
tensorboard==2.11.2
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.11.0
tensorflow-estimator==2.11.0
tensorflow-intel==2.11.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
tifffile==2021.11.2
tqdm==4.66.1
Twisted==23.8.0
twisted-iocpsupport==1.0.4
txaio==23.1.1
typing_extensions==4.7.1
urllib3==2.0.7
vidgear==0.3.1
Werkzeug==2.2.3
wincertstore==0.2
wrapt==1.16.0
zipp==3.15.0
zope.interface==6.4
